# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  user.ChangePasswordRequest:
    properties:
      confirm_password:
        type: string
      new_password:
        maxLength: 30
        minLength: 6
        type: string
      old_password:
        maxLength: 30
        minLength: 6
        type: string
    required:
    - confirm_password
    - new_password
    - old_password
    type: object
  user.ChangeRoleRequest:
    properties:
      new_role:
        $ref: '#/definitions/user.Role'
    required:
    - new_role
    type: object
  user.CreateUserRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 6
        type: string
      role:
        $ref: '#/definitions/user.Role'
    required:
    - email
    - first_name
    - last_name
    - password
    - role
    type: object
  user.Role:
    enum:
    - USER
    - MANAGER
    - ADMIN
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleManager
    - RoleAdmin
host: localhost:8080
info:
  contact: {}
  description: Backend API cho hệ thống smart parking
  title: Smart Parking Backend API
  version: "1.0"
paths:
...
  /slot-histories/{slotId}:
    get:
      description: Trả về lịch sử hoạt động của một vị trí đỗ xe theo slot ID
      parameters:
      - description: ID vị trí đỗ xe
        in: path
        name: slotId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Lấy lịch sử theo vị trí đỗ
      tags:
      - slot_history
  /users:
    get:
      description: Lấy danh sách người dùng có phân trang và tìm kiếm
      parameters:
      - default: 1
        description: Trang hiện tại
        in: query
        name: page
        type: integer
      - default: 10
        description: Số lượng mỗi trang
        in: query
        name: limit
        type: integer
      - description: Từ khóa tìm kiếm
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Lấy danh sách người dùng
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Tạo mới người dùng bởi quản trị viên
      parameters:
      - description: Thông tin người dùng
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Admin tạo người dùng
      tags:
      - user
  /users/change-password:
    patch:
      consumes:
      - application/json
      description: Người dùng hiện tại đổi mật khẩu của chính mình
      parameters:
      - description: Thông tin đổi mật khẩu
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Đổi mật khẩu
      tags:
      - user
  /users/change-role/{id}:
    patch:
      consumes:
      - application/json
      description: Quản trị viên đổi vai trò của người dùng theo ID
      parameters:
      - description: ID người dùng
        in: path
        name: id
        required: true
        type: integer
      - description: Vai trò mới
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.ChangeRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Đổi vai trò người dùng
      tags:
      - user
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"